% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\name{testEnrichment}
\alias{testEnrichment}
\title{Simple test for enrichment in a given set of genes}
\usage{
testEnrichment(module, gsets, background)
}
\arguments{
\item{module}{character vector containing the genes tested for enrichment}

\item{gsets}{a list containing lists of genes}

\item{background}{character vector}
}
\value{
Dataframe containing for every gene set which passed the filtering: \itemize{
 \item p-value of enrichment
 \item q-value of enrichment (p-value corrected for multiple testing)
 \item estimated odds score of enrichment, how much more likely is a given gene set to be part of
}
}
\description{
Assumes gsets is already filtered for background, will not check this!
}

